# docs/openapi/spec.yaml
openapi: 3.1.0
info:
  title: Eazy Recycling Backend Service
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /materials/{id}:
    get:
      tags:
      - material-controller
      operationId: getMaterialById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MaterialResponse"
    put:
      tags:
      - material-controller
      operationId: updateMaterial
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MaterialResponse"
    delete:
      tags:
      - material-controller
      operationId: deleteMaterial
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
  /material-prices/{id}:
    get:
      tags:
      - material-price-controller
      operationId: getMaterialPrice
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MaterialPriceResponse"
    put:
      tags:
      - material-price-controller
      operationId: updateMaterialPrice
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialPriceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MaterialPriceResponse"
    post:
      tags:
      - material-price-controller
      operationId: createMaterialPrice
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialPriceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MaterialPriceResponse"
    delete:
      tags:
      - material-price-controller
      operationId: deleteMaterialPrice
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
  /material-groups/{id}:
    get:
      tags:
      - material-group-controller
      operationId: getMaterialGroupById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MaterialGroupResponse"
    put:
      tags:
      - material-group-controller
      operationId: updateMaterialGroup
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialGroupRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MaterialGroupResponse"
    delete:
      tags:
      - material-group-controller
      operationId: deleteMaterialGroup
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
...
  /materials:
    get:
      tags:
      - material-controller
      operationId: getAllMaterials
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaterialResponse"
    post:
      tags:
      - material-controller
      operationId: createMaterial
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MaterialResponse"
  /material-groups:
    get:
      tags:
      - material-group-controller
      operationId: getAllMaterialGroups
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaterialGroupResponse"
    post:
      tags:
      - material-group-controller
      operationId: createMaterialGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialGroupRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MaterialGroupResponse"
...
  /catalog/items:
    get:
      tags:
      - catalog-controller
      operationId: searchCatalogItems
      parameters:
      - name: query
        in: query
        required: false
        schema:
          type: string
      - name: consignorPartyId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogItemResponse"
...
components:
  schemas:
...
    MaterialRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        materialGroupId:
          type: string
          format: uuid
        unitOfMeasure:
          type: string
          minLength: 1
        vatCode:
          type: string
          minLength: 1
        salesAccountNumber:
          type: string
        purchaseAccountNumber:
          type: string
        status:
          type: string
          minLength: 1
      required:
      - code
      - materialGroupId
      - name
      - status
      - unitOfMeasure
      - vatCode
    MaterialResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        materialGroupId:
          type: string
          format: uuid
        materialGroupCode:
          type: string
        materialGroupName:
          type: string
        unitOfMeasure:
          type: string
        vatCode:
          type: string
        salesAccountNumber:
          type: string
        purchaseAccountNumber:
          type: string
        defaultPrice:
          type: number
        status:
          type: string
        createdAt:
          type: string
        createdByName:
          type: string
        updatedAt:
          type: string
        updatedByName:
          type: string
      required:
      - code
      - id
      - name
      - status
      - unitOfMeasure
      - vatCode
    MaterialPriceRequest:
      type: object
      properties:
        price:
          type: number
      required:
      - price
    MaterialPriceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        materialGroupId:
          type: string
          format: uuid
        materialGroupName:
          type: string
        unitOfMeasure:
          type: string
        defaultPrice:
          type: number
      required:
      - code
      - id
      - name
      - unitOfMeasure
    MaterialGroupRequest:
      type: object
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        description:
          type: string
      required:
      - code
      - name
    MaterialGroupResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
        createdByName:
          type: string
        updatedAt:
          type: string
        updatedByName:
          type: string
      required:
      - code
      - id
      - name
...
    CatalogItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        itemType:
          type: string
          enum:
          - MATERIAL
          - PRODUCT
          - WASTE_STREAM
        code:
          type: string
        name:
          type: string
        unitOfMeasure:
          type: string
        vatCode:
          type: string
        categoryName:
          type: string
        consignorPartyId:
          type: string
          format: uuid
        purchaseAccountNumber:
          type: string
        salesAccountNumber:
          type: string
        wasteStreamNumber:
          type: string
        defaultPrice:
          type: number
      required:
      - code
      - id
      - itemType
      - name
      - unitOfMeasure
      - vatCode
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT